combined_requirements.txt
# Combined requirements for geometric thing and facenet-pytorch
# Using newer PyTorch versions from facenet-pytorch for compatibility

# Core dependencies from facenet-pytorch
numpy>=1.24.0,<2.0.0
Pillow>=10.2.0,<10.3.0
requests>=2.0.0,<3.0.0
torch>=2.2.0,<=2.3.0
torchvision>=0.17.0,<=0.18.0
tqdm>=4.0.0,<5.0.0

# Additional dependencies from geometric thing
matplotlib>=3.3.4
h5py>=2.10.0
imageio>=2.8.0
scipy
scikit_learn_extra>=0.2.0
scikit_learn
